Refactor: extract chat functionality into ChatBox component and streamline ChatPage logic
